[scanner] refactor: split MCP types, helm, networking, compute, and storage into focused modules
Fixes #18908
Fixes #18910
Fixes #18911
