+++ web/src/hooks/mcp/compute/index.ts
@@ -1,0 +1,4 @@
+export * from './useNodes'
+export * from './useGPU'
+export * from './useResourceQuotas'
+export { __computeTestables } from './gpuShared'
