+++ web/src/hooks/mcp/helm/useHelmReleases.ts
@@ -1,0 +1,31 @@
+import {
+  getDemoHelmHistory,
+  getDemoHelmReleases,
+  getDemoHelmValues,
+  HELM_CACHE_TTL_MS,
+  HELM_HISTORY_CACHE_KEY,
+  HELM_REFRESH_INTERVAL_MS,
+  HELM_RELEASES_CACHE_KEY,
+  loadHelmHistoryFromStorage,
+  loadHelmReleasesFromStorage,
+  saveHelmHistoryToStorage,
+  saveHelmReleasesToStorage,
+} from './shared'
+
+export { useHelmReleases } from './useHelmReleasesCore'
+export { useHelmHistory } from './useHelmHistoryCore'
+export { useHelmValues } from './useHelmValuesCore'
+
+export const __helmTestables = {
+  getDemoHelmReleases,
+  getDemoHelmHistory,
+  getDemoHelmValues,
+  loadHelmReleasesFromStorage,
+  saveHelmReleasesToStorage,
+  loadHelmHistoryFromStorage,
+  saveHelmHistoryToStorage,
+  HELM_RELEASES_CACHE_KEY,
+  HELM_HISTORY_CACHE_KEY,
+  HELM_CACHE_TTL_MS,
+  HELM_REFRESH_INTERVAL_MS,
+}
