+++ web/src/hooks/mcp/networking/useIngresses.ts
@@ -1,0 +1,1 @@
+export { useIngresses } from './ingressesCore'
