+++ web/src/hooks/mcp/networking/useNetworkPolicies.ts
@@ -1,0 +1,1 @@
+export { useNetworkPolicies } from './networkPoliciesCore'
