+++ web/src/hooks/mcp/networking/useServices.ts
@@ -1,0 +1,9 @@
+import { getDemoServices, loadServicesCacheFromStorage } from './shared'
+
+export { useServices } from './servicesCore'
+export { subscribeNetworkingCache } from './shared'
+
+export const __networkingTestables = {
+  loadServicesCacheFromStorage,
+  getDemoServices,
+}
