+++ web/src/hooks/mcp/storage.ts
@@ -1,803 +1,1 @@
-import { useState, useEffect, useCallback, useRef } from 'react'
-import { reportAgentDataSuccess, isAgentUnavailable } from '../useLocalAgent'
-import { isDemoMode } from '../../lib/demoMode'
-import { registerCacheReset, registerRefetch } from '../../lib/modeTransition'
-import { kubectlProxy } from '../../lib/kubectlProxy'
-import { REFRESH_INTERVAL_MS, getEffectiveInterval, getLocalAgentURL, agentFetch, clusterCacheRef } from './shared'
-import { deduplicateClustersByServer } from './dedup'
-import { subscribePolling } from './pollingManager'
-import { settledWithConcurrency } from '../../lib/utils/concurrency'
-import { MCP_HOOK_TIMEOUT_MS, LOCAL_AGENT_HTTP_URL, FETCH_DEFAULT_TIMEOUT_MS } from '../../lib/constants/network'
-import { CONSECUTIVE_FAILURE_THRESHOLD } from '../../lib/cache'
-import { isClusterModeBackend } from '../../lib/cache/fetcherUtils'
-import { useClusterResourceQuery } from './useClusterResourceQuery'
-import type { PVC, PV, ResourceQuota, LimitRange, ResourceQuotaSpec } from './types'
-
-// ---------------------------------------------------------------------------
-// Shared Storage State - enables cache reset notifications to all consumers
-// ---------------------------------------------------------------------------
-
-interface StorageSharedState {
-  cacheVersion: number
-  isResetting: boolean
-}
-
-let storageSharedState: StorageSharedState = {
-  cacheVersion: 0,
-  isResetting: false,
-}
-
-type StorageSubscriber = (state: StorageSharedState) => void
-const storageSubscribers = new Set<StorageSubscriber>()
-
-function notifyStorageSubscribers() {
-  Array.from(storageSubscribers).forEach(subscriber => subscriber(storageSharedState))
-}
-
-export function subscribeStorageCache(callback: StorageSubscriber): () => void {
-  storageSubscribers.add(callback)
-  return () => storageSubscribers.delete(callback)
-}
-
-// Module-level cache for PVCs data (persists across navigation)
-const PVCS_CACHE_KEY = 'kubestellar-pvcs-cache'
-
-interface PVCsCache {
-  data: PVC[]
-  timestamp: Date
-  key: string
-}
-
-let pvcsCache: PVCsCache | null = null
-
-// Load PVCs cache from localStorage
-function loadPVCsCacheFromStorage(cacheKey: string): { data: PVC[], timestamp: Date } | null {
-  try {
-    const stored = localStorage.getItem(PVCS_CACHE_KEY)
-    if (stored) {
-      const parsed = JSON.parse(stored)
-      if (parsed.key === cacheKey && Array.isArray(parsed.data) && parsed.data.length > 0) {
-        const timestamp = parsed.timestamp ? new Date(parsed.timestamp) : new Date()
-        pvcsCache = { data: parsed.data, timestamp, key: cacheKey }
-        return { data: parsed.data, timestamp }
-      }
-    }
-  } catch {
-    // Ignore parse errors
-  }
-  return null
-}
-
-function savePVCsCacheToStorage() {
-  if (pvcsCache) {
-    try {
-      localStorage.setItem(PVCS_CACHE_KEY, JSON.stringify({
-        data: pvcsCache.data,
-        timestamp: pvcsCache.timestamp.toISOString(),
-        key: pvcsCache.key
-      }))
-    } catch {
-      // Ignore storage errors
-    }
-  }
-}
-
-// Hook to get PVCs with localStorage-backed caching
-export function usePVCs(cluster?: string, namespace?: string) {
-  const cacheKey = `pvcs:${cluster || 'all'}:${namespace || 'all'}`
-
-  // Initialize from cache if available
-  const getCachedData = () => {
-    if (pvcsCache && pvcsCache.key === cacheKey) {
-      return { data: pvcsCache.data, timestamp: pvcsCache.timestamp }
-    }
-    return loadPVCsCacheFromStorage(cacheKey)
-  }
-
-  const cached = getCachedData()
-  const [pvcs, setPVCs] = useState<PVC[]>(cached?.data || [])
-  const [isLoading, setIsLoading] = useState(!cached)
-  const [isRefreshing, setIsRefreshing] = useState(false)
-  const [lastUpdated, setLastUpdated] = useState<Date | null>(cached?.timestamp || null)
-  const [error, setError] = useState<string | null>(null)
-  const [consecutiveFailures, setConsecutiveFailures] = useState(0)
-  const [lastRefresh, setLastRefresh] = useState<Date | null>(cached?.timestamp || null)
-
-  // Track mounted state to prevent state updates after unmount (StrictMode)
-  const isMountedRef = useRef(true)
-  useEffect(() => {
-    isMountedRef.current = true
-    return () => { isMountedRef.current = false }
-  }, [])
-
-  // Reset state when cluster changes (only if still mounted)
-  // Don't reset to loading if we have cached data (stale-while-revalidate)
-  useEffect(() => {
-    if (!isMountedRef.current) return
-    const newCacheKey = `pvcs:${cluster || 'all'}:${namespace || 'all'}`
-    const hasCached = pvcsCache && pvcsCache.key === newCacheKey
-    if (!hasCached) {
-      setPVCs([])
-      setIsLoading(true)
-    }
-    setError(null)
-  }, [cluster, namespace])
-
-  const refetch = useCallback(async (silent = false) => {
-    if (!isMountedRef.current) return
-    if (!silent) {
-      setIsRefreshing(true)
-    }
-    // If demo mode is enabled, use demo data
-    if (isDemoMode()) {
-      const demoPVCs = getDemoPVCs().filter(p =>
-        (!cluster || p.cluster === cluster) && (!namespace || p.namespace === namespace)
-      )
-      if (!isMountedRef.current) return
-      setPVCs(demoPVCs)
-      setIsLoading(false)
-      setIsRefreshing(false)
-      setError(null)
-      setLastUpdated(new Date())
-      return
-    }
-
-    // Try local agent HTTP endpoint first
-    if (!isAgentUnavailable()) {
-      try {
-        // If cluster is specified, fetch from that cluster only
-        // If no cluster specified, aggregate from all clusters
-        const allClusters = clusterCacheRef.clusters.filter(c => c.reachable !== false)
-        const dedupClusters = deduplicateClustersByServer(allClusters)
-        const clustersToFetch = cluster
-          ? [{ name: cluster, context: cluster }]
-          : dedupClusters
-
-        if (clustersToFetch.length > 0) {
-          const allPVCs: PVC[] = []
-          let anySuccess = false
-
-          // Fetch PVCs from each cluster (in parallel for speed)
-          const fetchTasks = clustersToFetch.map((c) => async () => {
-            try {
-              const params = new URLSearchParams()
-              params.append('cluster', c.context || c.name)
-              if (namespace) params.append('namespace', namespace)
-              const controller = new AbortController()
-              const timeoutId = setTimeout(() => controller.abort(), MCP_HOOK_TIMEOUT_MS)
-              const response = await agentFetch(`${getLocalAgentURL()}/pvcs?${params}`, {
-                signal: controller.signal,
-                headers: { 'Accept': 'application/json' },
-              })
-              clearTimeout(timeoutId)
-              if (response.ok) {
-                const agentData = await response.json()
-                const mappedPVCs: PVC[] = (agentData.pvcs || []).map((p: PVC) => ({ ...p, cluster: c.name }))
-                return { success: true, pvcs: mappedPVCs }
-              }
-            } catch {
-              // Individual cluster failure - continue with others
-            }
-            return { success: false, pvcs: [] }
-          })
-
-          const settled = await settledWithConcurrency(fetchTasks)
-          for (const entry of (settled || [])) {
-            if (entry.status === 'fulfilled' && entry.value.success) {
-              anySuccess = true
-              allPVCs.push(...entry.value.pvcs)
-            }
-          }
-
-          if (anySuccess) {
-            const now = new Date()
-            pvcsCache = { data: allPVCs, timestamp: now, key: cacheKey }
-            savePVCsCacheToStorage()
-            if (!isMountedRef.current) return
-            setPVCs(allPVCs)
-            setError(null)
-            setLastUpdated(now)
-            setConsecutiveFailures(0)
-            setLastRefresh(now)
-            setIsLoading(false)
-            setIsRefreshing(false)
-            reportAgentDataSuccess()
-            return
-          }
-        }
-      } catch {
-        // Fall through to kubectl proxy
-      }
-    }
-
-    // Try kubectl proxy as fallback
-    if (!isAgentUnavailable()) {
-      try {
-        const allClusters = clusterCacheRef.clusters.filter(c => c.reachable !== false)
-        const dedupClusters = deduplicateClustersByServer(allClusters)
-        const clustersToFetch = cluster
-          ? [{ name: cluster, context: clusterCacheRef.clusters.find(c => c.name === cluster)?.context || cluster }]
-          : dedupClusters
-
-        if (clustersToFetch.length > 0) {
-          const allPVCs: PVC[] = []
-          let anySuccess = false
-
-          for (const c of (clustersToFetch || [])) {
-            try {
-              const kubectlContext = c.context || c.name
-              const pvcData = await kubectlProxy.getPVCs(kubectlContext, namespace)
-              const mappedPVCs: PVC[] = pvcData.map(p => ({
-                name: p.name,
-                namespace: p.namespace,
-                cluster: c.name,
-                status: p.status,
-                capacity: p.capacity,
-                storageClass: p.storageClass,
-              }))
-              allPVCs.push(...mappedPVCs)
-              anySuccess = true
-            } catch {
-              // Individual cluster failure - continue with others
-            }
-          }
-
-          if (anySuccess) {
-            const now = new Date()
-            pvcsCache = { data: allPVCs, timestamp: now, key: cacheKey }
-            savePVCsCacheToStorage()
-            if (!isMountedRef.current) return
-            setPVCs(allPVCs)
-            setError(null)
-            setLastUpdated(now)
-            setConsecutiveFailures(0)
-            setLastRefresh(now)
-            setIsLoading(false)
-            setIsRefreshing(false)
-            return
-          }
-        }
-      } catch {
-        console.error(`[usePVCs] kubectl proxy failed, trying API`)
-      }
-    }
-
-    if (!isMountedRef.current) return
-    if (!silent) {
-      const hasCachedData = pvcsCache && pvcsCache.key === cacheKey
-      if (!hasCachedData) {
-        setIsLoading(true)
-      }
-    }
-    try {
-      const params = new URLSearchParams()
-      if (cluster) params.append('cluster', cluster)
-      if (namespace) params.append('namespace', namespace)
-      if (isClusterModeBackend()) {
-        try {
-          const response = await fetch(`/api/mcp/pvcs?${params}`, {
-            signal: AbortSignal.timeout(MCP_HOOK_TIMEOUT_MS),
-          })
-          if (response.ok) {
-            const data = await response.json()
-            const newData = data.pvcs || []
-            const now = new Date()
-
-            pvcsCache = { data: newData, timestamp: now, key: cacheKey }
-            savePVCsCacheToStorage()
-
-            if (!isMountedRef.current) return
-            setPVCs(newData)
-            setError(null)
-            setLastUpdated(now)
-            setConsecutiveFailures(0)
-            setLastRefresh(now)
-            setIsLoading(false)
-            return
-          }
-        } catch (err) {
-          console.error('[pvcs] Backend fetch failed:', err)
-          // Error propagated via hook error state; log here for debugging
-        }
-        if (!isMountedRef.current) return
-        setIsLoading(false)
-        return
-      }
-      const resp = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/pvcs?${params}`)
-      if (!resp.ok) throw new Error(`HTTP ${resp.status}`)
-      const data = await resp.json()
-      const newData = data.pvcs || []
-      const now = new Date()
-
-      // Update module-level cache
-      pvcsCache = { data: newData, timestamp: now, key: cacheKey }
-      savePVCsCacheToStorage()
-
-      if (!isMountedRef.current) return
-      setPVCs(newData)
-      setError(null)
-      setLastUpdated(now)
-      setConsecutiveFailures(0)
-      setLastRefresh(now)
-    } catch (err: unknown) {
-      if (!isMountedRef.current) return
-      const message = err instanceof Error ? err.message : 'Failed to fetch PVCs'
-      setConsecutiveFailures(prev => prev + 1)
-      setLastRefresh(new Date())
-      if (!silent && !pvcsCache) {
-        setError(message)
-      }
-    } finally {
-      if (isMountedRef.current) {
-        if (!silent) {
-          setIsLoading(false)
-        }
-        setIsRefreshing(false)
-      }
-    }
-  }, [cluster, namespace, cacheKey])
-
-  useEffect(() => {
-    let cancelled = false
-
-    const doFetch = async () => {
-      const hasCachedData = pvcsCache && pvcsCache.key === cacheKey
-      if (!cancelled) {
-        await refetch(!!hasCachedData)
-      }
-    }
-
-    void doFetch()
-
-    const unregisterRefetch = registerRefetch(`pvcs:${cacheKey}`, () => {
-      if (!cancelled) {
-        void refetch(false)
-      }
-    })
-
-    return () => {
-      cancelled = true
-      unregisterRefetch()
-    }
-  }, [cacheKey, refetch])
-
-  useEffect(() => {
-    let cancelled = false
-
-    const unsubscribePolling = subscribePolling(
-      `pvcs:${cacheKey}`,
-      getEffectiveInterval(REFRESH_INTERVAL_MS, consecutiveFailures),
-      () => {
-        if (!cancelled) {
-          void refetch(true)
-        }
-      },
-    )
-
-    return () => {
-      cancelled = true
-      unsubscribePolling()
-    }
-  }, [cacheKey, consecutiveFailures, refetch])
-
-  // Subscribe to cache reset notifications - triggers skeleton when cache is cleared
-  useEffect(() => {
-    const handleCacheReset = (state: StorageSharedState) => {
-      if (state.isResetting) {
-        setIsLoading(true)
-        setPVCs([])
-        setLastUpdated(null)
-      }
-    }
-    return subscribeStorageCache(handleCacheReset)
-  }, [])
-
-  return {
-    pvcs,
-    isLoading,
-    isRefreshing,
-    lastUpdated,
-    error,
-    refetch: () => refetch(false),
-    consecutiveFailures,
-    isFailed: consecutiveFailures >= CONSECUTIVE_FAILURE_THRESHOLD,
-    lastRefresh,
-  }
-}
-
-// Hook to get PVs (PersistentVolumes)
-export function usePVs(cluster?: string) {
-  const [pvs, setPVs] = useState<PV[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [isRefreshing, setIsRefreshing] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [consecutiveFailures, setConsecutiveFailures] = useState(0)
-
-  // Track mounted state to prevent state updates after unmount (StrictMode)
-  const isMountedRef = useRef(true)
-  useEffect(() => {
-    isMountedRef.current = true
-    return () => { isMountedRef.current = false }
-  }, [])
-
-  const refetch = useCallback(async () => {
-    if (!isMountedRef.current) return
-    setIsRefreshing(true)
-
-    // If demo mode is enabled, use demo data
-    if (isDemoMode()) {
-      if (!isMountedRef.current) return
-      setPVs([])
-      setIsLoading(false)
-      setIsRefreshing(false)
-      setError(null)
-      return
-    }
-
-    if (isAgentUnavailable()) {
-      if (isMountedRef.current) {
-        setError('Agent unavailable')
-        setConsecutiveFailures(prev => prev + 1)
-        setIsLoading(false)
-        setIsRefreshing(false)
-      }
-      return
-    }
-
-    try {
-      const allClusters = clusterCacheRef.clusters.filter(c => c.reachable !== false)
-      const dedupClusters = deduplicateClustersByServer(allClusters)
-      const clustersToFetch = cluster
-        ? [{ name: cluster, context: cluster }]
-        : dedupClusters
-
-      if (clustersToFetch.length === 0) {
-        if (isMountedRef.current) {
-          setPVs([])
-          setIsLoading(false)
-          setIsRefreshing(false)
-          setError(null)
-        }
-        return
-      }
-
-      const allPVs: PV[] = []
-      let anySuccess = false
-
-      const fetchTasks = clustersToFetch.map((c) => async () => {
-        try {
-          const params = new URLSearchParams()
-          params.append('cluster', c.context || c.name)
-          if (isClusterModeBackend()) {
-            try {
-              const response = await fetch(`/api/mcp/pvs?${params}`, {
-                signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS),
-              })
-              if (response.ok) {
-                const agentData = await response.json()
-                const mappedPVs: PV[] = (agentData.pvs || []).map((p: PV) => ({ ...p, cluster: c.name }))
-                return { success: true, pvs: mappedPVs }
-              }
-            } catch (err) {
-              console.error('[pvs] Backend fetch failed:', err)
-              // Error propagated; caller handles fallback to agent fetch
-            }
-            return { success: false, pvs: [] }
-          }
-          const controller = new AbortController()
-          const timeoutId = setTimeout(() => controller.abort(), MCP_HOOK_TIMEOUT_MS)
-          const response = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/pvs?${params}`, {
-            signal: controller.signal,
-            headers: { 'Accept': 'application/json' },
-          })
-          clearTimeout(timeoutId)
-          if (response.ok) {
-            const agentData = await response.json()
-            const mappedPVs: PV[] = (agentData.pvs || []).map((p: PV) => ({ ...p, cluster: c.name }))
-            return { success: true, pvs: mappedPVs }
-          }
-        } catch {
-          // Individual cluster failure - continue with others
-        }
-        return { success: false, pvs: [] }
-      })
-
-      const settled = await settledWithConcurrency(fetchTasks)
-      for (const entry of (settled || [])) {
-        if (entry.status === 'fulfilled' && entry.value.success) {
-          anySuccess = true
-          allPVs.push(...entry.value.pvs)
-        }
-      }
-
-      if (!isMountedRef.current) return
-      if (anySuccess) {
-        setPVs(allPVs)
-        setError(null)
-        setConsecutiveFailures(0)
-      } else {
-        setError('Failed to fetch PVs from any cluster')
-        setConsecutiveFailures(prev => prev + 1)
-      }
-    } catch (err: unknown) {
-      if (isMountedRef.current) {
-        const message = err instanceof Error ? err.message : 'Failed to fetch PVs'
-        setError(message)
-        setConsecutiveFailures(prev => prev + 1)
-      }
-    } finally {
-      if (isMountedRef.current) {
-        setIsLoading(false)
-        setIsRefreshing(false)
-      }
-    }
-  }, [cluster])
-
-  useEffect(() => {
-    void refetch()
-    return registerRefetch(`pvs:${cluster || 'all'}`, () => {
-      void refetch()
-    })
-  }, [cluster, refetch])
-
-  useEffect(() => {
-    return subscribePolling(
-      `pvs:${cluster || 'all'}`,
-      getEffectiveInterval(REFRESH_INTERVAL_MS, consecutiveFailures),
-      () => {
-        void refetch()
-      },
-    )
-  }, [cluster, consecutiveFailures, refetch])
-
-  return { pvs, isLoading, isRefreshing, error, refetch, consecutiveFailures, isFailed: consecutiveFailures >= CONSECUTIVE_FAILURE_THRESHOLD }
-}
-
-// Hook to get ResourceQuotas
-// When forceLive is true, skip demo mode fallback and always query the real API.
-// Used by GPU Reservations to show live data when running in-cluster with OAuth.
-// Returns `isDemoFallback: true` when the hook is serving demo data so callers
-// can render the Demo badge only for true demo output. See Issue 9356.
-export function useResourceQuotas(cluster?: string, namespace?: string, forceLive = false) {
-  const result = useClusterResourceQuery<ResourceQuota>({
-    resourceKey: 'resourceQuotas',
-    endpoint: 'resourcequotas',
-    dataField: 'resourceQuotas',
-    getDemoData: getDemoResourceQuotas,
-    filterFn: (item, c, ns) => (!c || item.cluster === c) && (!ns || item.namespace === ns),
-    cluster,
-    namespace,
-    forceLive,
-    silentErrors: true,
-  })
-
-  return {
-    resourceQuotas: result.data,
-    isLoading: result.isLoading,
-    error: result.error,
-    refetch: result.refetch,
-    isDemoFallback: result.isDemoFallback,
-  }
-}
-
-// Hook to get LimitRanges
-export function useLimitRanges(cluster?: string, namespace?: string) {
-  const result = useClusterResourceQuery<LimitRange>({
-    resourceKey: 'limitRanges',
-    endpoint: 'limitranges',
-    dataField: 'limitRanges',
-    getDemoData: getDemoLimitRanges,
-    filterFn: (item, c, ns) => (!c || item.cluster === c) && (!ns || item.namespace === ns),
-    cluster,
-    namespace,
-    silentErrors: true,
-  })
-
-  return {
-    limitRanges: result.data,
-    isLoading: result.isLoading,
-    error: result.error,
-    refetch: result.refetch,
-  }
-}
-
-// Create or update a ResourceQuota
-export async function createOrUpdateResourceQuota(spec: ResourceQuotaSpec): Promise<ResourceQuota> {
-  if (isClusterModeBackend()) {
-    const response = await fetch('/api/mcp/resourcequotas', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(spec),
-      signal: AbortSignal.timeout(MCP_HOOK_TIMEOUT_MS),
-    })
-    if (!response.ok) throw new Error(`HTTP ${response.status}`)
-    const data = await response.json()
-    return data.resourceQuota
-  }
-
-  const resp = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/resourcequotas`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(spec),
-  })
-  if (!resp.ok) throw new Error(`HTTP ${resp.status}`)
-  const data = await resp.json()
-  return data.resourceQuota
-}
-
-// Delete a ResourceQuota
-export async function deleteResourceQuota(cluster: string, namespace: string, name: string): Promise<void> {
-  const params = new URLSearchParams({ cluster, namespace, name })
-
-  if (isClusterModeBackend()) {
-    const response = await fetch(`/api/mcp/resourcequotas?${params.toString()}`, {
-      method: 'DELETE',
-      signal: AbortSignal.timeout(MCP_HOOK_TIMEOUT_MS),
-    })
-    if (!response.ok) throw new Error(`HTTP ${response.status}`)
-    return
-  }
-
-  const resp = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/resourcequotas?${params.toString()}`, {
-    method: 'DELETE',
-  })
-  if (!resp.ok) throw new Error(`HTTP ${resp.status}`)
-}
-
-// Common GPU resource types for quotas
-export const GPU_RESOURCE_TYPES = [
-  { key: 'requests.nvidia.com/gpu', label: 'NVIDIA GPU Requests', description: 'Maximum GPUs that can be requested' },
-  { key: 'limits.nvidia.com/gpu', label: 'NVIDIA GPU Limits', description: 'Maximum GPU limits allowed' },
-  { key: 'requests.amd.com/gpu', label: 'AMD GPU Requests', description: 'Maximum AMD GPUs that can be requested' },
-  { key: 'limits.amd.com/gpu', label: 'AMD GPU Limits', description: 'Maximum AMD GPU limits allowed' },
-] as const
-
-// Common resource types for quotas
-export const COMMON_RESOURCE_TYPES = [
-  { key: 'requests.cpu', label: 'CPU Requests', description: 'Total CPU requests allowed' },
-  { key: 'limits.cpu', label: 'CPU Limits', description: 'Total CPU limits allowed' },
-  { key: 'requests.memory', label: 'Memory Requests', description: 'Total memory requests allowed' },
-  { key: 'limits.memory', label: 'Memory Limits', description: 'Total memory limits allowed' },
-  { key: 'pods', label: 'Pods', description: 'Maximum number of pods' },
-  { key: 'services', label: 'Services', description: 'Maximum number of services' },
-  { key: 'persistentvolumeclaims', label: 'PVCs', description: 'Maximum number of PVCs' },
-  { key: 'requests.storage', label: 'Storage Requests', description: 'Total storage that can be requested' },
-  ...GPU_RESOURCE_TYPES,
-] as const
-
-// Demo data functions (not exported)
-
-function getDemoPVCs(): PVC[] {
-  return [
-    { name: 'postgres-data', namespace: 'data', cluster: 'prod-east', status: 'Bound', storageClass: 'gp3', capacity: '100Gi', accessModes: ['ReadWriteOnce'], volumeName: 'pvc-abc123', age: '40d' },
-    { name: 'redis-data', namespace: 'data', cluster: 'prod-east', status: 'Bound', storageClass: 'gp3', capacity: '20Gi', accessModes: ['ReadWriteOnce'], volumeName: 'pvc-def456', age: '40d' },
-    { name: 'prometheus-data', namespace: 'monitoring', cluster: 'staging', status: 'Bound', storageClass: 'standard', capacity: '50Gi', accessModes: ['ReadWriteOnce'], volumeName: 'pvc-ghi789', age: '20d' },
-    { name: 'grafana-data', namespace: 'monitoring', cluster: 'staging', status: 'Bound', storageClass: 'standard', capacity: '10Gi', accessModes: ['ReadWriteOnce'], volumeName: 'pvc-jkl012', age: '20d' },
-    { name: 'model-cache', namespace: 'ml', cluster: 'vllm-d', status: 'Bound', storageClass: 'fast-ssd', capacity: '500Gi', accessModes: ['ReadWriteMany'], volumeName: 'pvc-mno345', age: '15d' },
-    { name: 'training-data', namespace: 'ml', cluster: 'vllm-d', status: 'Pending', storageClass: 'fast-ssd', capacity: '1Ti', accessModes: ['ReadWriteMany'], age: '1d' },
-    { name: 'logs-archive', namespace: 'logging', cluster: 'prod-east', status: 'Bound', storageClass: 'cold-storage', capacity: '200Gi', accessModes: ['ReadWriteOnce'], volumeName: 'pvc-pqr678', age: '60d' },
-  ]
-}
-
-function getDemoResourceQuotas(): ResourceQuota[] {
-  return [
-    {
-      name: 'compute-quota',
-      namespace: 'production',
-      cluster: 'prod-east',
-      hard: { 'requests.cpu': '10', 'requests.memory': '20Gi', 'limits.cpu': '20', 'limits.memory': '40Gi', pods: '50' },
-      used: { 'requests.cpu': '5', 'requests.memory': '10Gi', 'limits.cpu': '8', 'limits.memory': '16Gi', pods: '25' },
-      age: '30d'
-    },
-    {
-      name: 'storage-quota',
-      namespace: 'data',
-      cluster: 'prod-east',
-      hard: { 'requests.storage': '500Gi', persistentvolumeclaims: '10' },
-      used: { 'requests.storage': '320Gi', persistentvolumeclaims: '5' },
-      age: '40d'
-    },
-    {
-      name: 'ml-quota',
-      namespace: 'ml',
-      cluster: 'vllm-d',
-      hard: { 'requests.cpu': '100', 'requests.memory': '200Gi', 'limits.cpu': '200', 'limits.memory': '400Gi', 'requests.nvidia.com/gpu': '8', pods: '20' },
-      used: { 'requests.cpu': '64', 'requests.memory': '128Gi', 'limits.cpu': '128', 'limits.memory': '256Gi', 'requests.nvidia.com/gpu': '4', pods: '8' },
-      age: '15d'
-    },
-    {
-      name: 'default-quota',
-      namespace: 'default',
-      cluster: 'staging',
-      hard: { 'requests.cpu': '4', 'requests.memory': '8Gi', 'limits.cpu': '8', 'limits.memory': '16Gi', pods: '20' },
-      used: { 'requests.cpu': '1', 'requests.memory': '2Gi', 'limits.cpu': '2', 'limits.memory': '4Gi', pods: '5' },
-      age: '60d'
-    },
-  ]
-}
-
-function getDemoLimitRanges(): LimitRange[] {
-  return [
-    {
-      name: 'container-limits',
-      namespace: 'production',
-      cluster: 'prod-east',
-      limits: [
-        {
-          type: 'Container',
-          default: { cpu: '500m', memory: '512Mi' },
-          defaultRequest: { cpu: '100m', memory: '128Mi' },
-          max: { cpu: '2', memory: '4Gi' },
-          min: { cpu: '50m', memory: '64Mi' }
-        }
-      ],
-      age: '30d'
-    },
-    {
-      name: 'pod-limits',
-      namespace: 'ml',
-      cluster: 'vllm-d',
-      limits: [
-        {
-          type: 'Container',
-          default: { cpu: '1', memory: '2Gi' },
-          defaultRequest: { cpu: '500m', memory: '1Gi' },
-          max: { cpu: '16', memory: '64Gi' },
-          min: { cpu: '100m', memory: '256Mi' }
-        },
-        {
-          type: 'Pod',
-          max: { cpu: '32', memory: '128Gi' }
-        }
-      ],
-      age: '15d'
-    },
-    {
-      name: 'storage-limits',
-      namespace: 'data',
-      cluster: 'prod-east',
-      limits: [
-        {
-          type: 'PersistentVolumeClaim',
-          max: { storage: '100Gi' },
-          min: { storage: '1Gi' }
-        }
-      ],
-      age: '40d'
-    },
-  ]
-}
-
-// Register with mode transition coordinator for unified cache clearing
-if (typeof window !== 'undefined') {
-  registerCacheReset('storage', () => {
-    // Set resetting flag to trigger skeleton display
-    storageSharedState = {
-      cacheVersion: storageSharedState.cacheVersion + 1,
-      isResetting: true,
-    }
-    notifyStorageSubscribers()
-
-    try {
-      localStorage.removeItem(PVCS_CACHE_KEY)
-    } catch {
-      // Ignore storage errors
-    }
-    pvcsCache = null
-
-    // Reset the resetting flag after a tick
-    setTimeout(() => {
-      storageSharedState = { ...storageSharedState, isResetting: false }
-      notifyStorageSubscribers()
-    }, 0)
-  })
-}
-
-export const __storageTestables = {
-  getDemoPVCs,
-  getDemoResourceQuotas,
-  getDemoLimitRanges,
-  loadPVCsCacheFromStorage,
-  savePVCsCacheToStorage,
-  PVCS_CACHE_KEY,
-}
+export * from './storage/index'
