+++ web/src/hooks/mcp/types/events.ts
@@ -1,0 +1,11 @@
+export interface ClusterEvent {
+  type: string
+  reason: string
+  message: string
+  object: string
+  namespace: string
+  cluster?: string
+  count: number
+  firstSeen?: string
+  lastSeen?: string
+}
