+++ web/src/hooks/mcp/types/helm.ts
@@ -1,0 +1,19 @@
+export interface HelmRelease {
+  name: string
+  namespace: string
+  revision: string
+  updated: string
+  status: string
+  chart: string
+  app_version: string
+  cluster?: string
+}
+
+export interface HelmHistoryEntry {
+  revision: number
+  updated: string
+  status: string
+  chart: string
+  app_version: string
+  description: string
+}
