+++ web/src/hooks/mcp/types/index.ts
@@ -1,0 +1,7 @@
+export * from './clusters'
+export * from './compute'
+export * from './storage'
+export * from './networking'
+export * from './helm'
+export * from './workloads'
+export * from './events'
