+++ web/src/hooks/mcp/types/storage.ts
@@ -1,0 +1,38 @@
+export interface PVC {
+  name: string
+  namespace: string
+  cluster?: string
+  status: string
+  storageClass?: string
+  capacity?: string
+  accessModes?: string[]
+  volumeName?: string
+  age?: string
+  labels?: Record<string, string>
+  annotations?: Record<string, string>
+}
+
+export interface PV {
+  name: string
+  cluster?: string
+  status: string
+  capacity?: string
+  storageClass?: string
+  reclaimPolicy?: string
+  accessModes?: string[]
+  claimRef?: string
+  volumeMode?: string
+  age?: string
+  labels?: Record<string, string>
+}
+
+export interface StorageClass {
+  name: string
+  cluster?: string
+  provisioner: string
+  reclaimPolicy?: string
+  volumeBindingMode?: string
+  isDefault?: boolean
+  age?: string
+  labels?: Record<string, string>
+}
